(refactor) Format redux store tree into exports
